fix: Prevent crash of ChannelModal
Move FocusTrap closer to content since at some point, the
content can be loading and can show a spinner. When showing
a spinner, we do not have any focusable element that can
be used for FocusTrap. By moving FocusTrap only around
the ModalFooter, we know precisely what is rendered below,
and we are sure that focusable buttons and the optional
terms checkbox are rendered