feat: Add `redirectAfterInstall` searchParam support in filtered list
Here we are in the list of connectors filtered on a category. For example in the app we have chosen a connector in the category "energy" and not directly "GEG". So we arrive in the Store on the list that displays GEG, Enercoop etc. It is then the Store that will set the value, depending on the connector we clicked on.