feat: Add helper for redirect to configure after konnector is installed
When we are already in an Intent,
we cannot use the `/configure` route,
because it ultimately returns the `IntentIframe` component
which creates a new `iframe` inside the first one.

This results in a `CSP:frame-ancestors` problem

It is therefore necessary to favor the use of
the `compose` service of `cozy-inter-app`
which will add the new `iframe` next to the first one.