feat: Update `INSTALL` Intent
This Intent can now redirect to 2 additional pages.
With the `category` option, we display the page of
konnectors/apps filtered by category.
With the `slug` and `activateDescriptionPageForSlug`
option, we display the description page of the konnector/app.

The code around the `redirectAfterInstall`
searchparams is no longer useful and can therefore be removed.

The rest of the process remains essentially identical
to the Intent `InstallAppIntent`(`INSTALL`)