feat(Intent): Add `terminateIfInstalled` option to intentData
Added an option to terminate the Intent when you want
to "Open" an installed but not connected connector.
Necessary because an app like Mespapiers needs to be
able to connect a connector already installed.

`disableClick` is no longer useful, because we want to
access the details of an already installed connector.